requirements.txt
colorama==0.4.6
filelock==3.13.1
fsspec==2024.6.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
pillow==11.1.0
setuptools==70.2.0
sympy==1.13.1
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
typing_extensions==4.12.2
